+++ graphite_api/functions.py
@@ -450,6 +450,28 @@ def keepLastValue(requestContext, seriesList, limit=INF):
     return seriesList
 
 
+def changed(requestContext, seriesList):
+    """
+    Takes one metric or a wildcard seriesList.
+    Output 1 when the value changed, 0 when null or the same
+    Example::
+
+        &target=changed(Server01.connections.handled)
+    """
+    for series in seriesList:
+        series.name = series.pathExpression = 'changed(%s)' % series.name
+        previous = None
+        for index, value in enumerate(series):
+            if previous is None:
+                series[index] = 0
+            elif value is not None and previous != value:
+                series[index] = 1
+            else:
+                series[index] = 0
+            previous = value
+    return seriesList
+
+
 def asPercent(requestContext, seriesList, total=None):
     """
 
@@ -3174,6 +3196,7 @@ SeriesFunctions = {
     'cumulative': cumulative,
     'consolidateBy': consolidateBy,
     'keepLastValue': keepLastValue,
+    'changed': changed,
     'drawAsInfinite': drawAsInfinite,
     'secondYAxis': secondYAxis,
     'lineWidth': lineWidth,