+++ docs/releases.rst
@@ -16,7 +16,7 @@ Graphite-API releases
   database-backed finders such as Cyanite because it allows fetching all time
   series at once instead of sequentially.
 
-* Add ``minimumBelow`` function.
+* Add ``minimumBelow`` and ``changed`` functions.
 
 * Fix parsing dates such as "feb27" during month days > 28.